modules/pub_sync/requirements.txt
structlog==21.5.0
environs==6.0.0
python-dateutil==2.8.1